refactor: Score 원시값 포장
