+++ src/main/java/chess/domain/piece/Bishop.java
@@ -5,6 +5,6 @@ public class Bishop extends Piece {
     private static final double SCORE = 3;
 
     public Bishop(TeamType teamType) {
-        super(NAME, teamType, SCORE, Direction.getBishopDirections());
+        super(NAME, teamType, new Score(SCORE), Direction.getBishopDirections());
     }
 }
