+++ src/main/java/chess/domain/piece/King.java
@@ -8,7 +8,7 @@ public class King extends Piece {
     private static final double SCORE = 0;
 
     public King(TeamType teamType) {
-        super(NAME, teamType, SCORE, Direction.getKingDirections());
+        super(NAME, teamType, new Score(SCORE), Direction.getKingDirections());
     }
 
     @Override
